refactor: extract requestdata UI to AlertRequestData component
